% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPerson.R
\name{checkPerson}
\alias{checkPerson}
\title{Check person dataframe}
\usage{
checkPerson(person)
}
\arguments{
\item{person}{person dataframe}
}
\value{
No return value, called for side effects
}
\description{
\code{checkPerson} checks that the person dataframe is formatted correctly
}
\examples{
library(LTASR)
library(dplyr)

#Import example peron file
person <- person_example \%>\%
  mutate(dob = as.Date(dob, format='\%m/\%d/\%Y'),
         pybegin = as.Date(pybegin, format='\%m/\%d/\%Y'),
         dlo = as.Date(dlo, format='\%m/\%d/\%Y'))

#Returns nothing
checkPerson(person)

}
