% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_history.R
\docType{data}
\name{history_example}
\alias{history_example}
\title{Example History File for Testing}
\format{
A data frame with 4 rows and 5 variables:
\describe{
\item{id}{unique identifier; numeric}
\item{begin_dt}{beginning date of an exposure period; character}
\item{end_dt}{beginning date of an exposure period; character}
\item{employed}{a hypothetical variable indicating employment during
the given exposure period; numeric (0/1)}
\item{exposure_level}{a hypothetical variable identifying daily exposure
levels to be summed to calculate a cumulative exposure; numeric}

...
}
}
\source{
{Internally Generated}
}
\usage{
history_example
}
\description{
A tibble containing example history file data to be used for testing and
demonstration of the package
}
\keyword{datasets}
