% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRate.R
\name{parseRate}
\alias{parseRate}
\title{Parses LTAS rate file in .xml format}
\usage{
parseRate(xmlpath)
}
\arguments{
\item{xmlpath}{path of LTAS rate file}
}
\value{
returns a list containing:
\enumerate{
\item $residual: the minor number where all unknown deaths will be assigned
\item $MinorDesc: a dataframe/tibble giving descriptions of minor numbers as well as
how minors are mapped to majors
\item $mapping: a dataframe/tibble listing how each icd-code and revision will be mapped
to each minor number
\item $age_cut: a numeric specifying cut-points for age strata
\item $cp_cut: a numeric specifying cut-points for calendar period strata
}
}
\description{
Parses LTAS rate file in .xml format
}
