% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_rate.R
\docType{data}
\name{us_119ucod_19602021}
\alias{us_119ucod_19602021}
\title{119 UCOD U.S. Death Rate, 1960-2021}
\format{
A list with 4 elements:
\describe{
\item{residual}{the minor/outcome number to which unknown/uncategorized outcomes will be mapped to}
\item{MinorDesc}{a data.frame containing descriptions for each minor and major grouping}
\item{mapping}{a tibble detailing which minor number each icd-code and revision combination will be mapped to}
\item{rates}{the population referent rate for each minor for each gender/race/calendar period/age strata}
...
}
}
\source{
{Available upon request from nioshltas@cdc.gov}
}
\usage{
us_119ucod_19602021
}
\description{
A list containing referent underlying cause of death (UCOD) rate information
for the US population from 1960-2021 for the 119 minor/outcome LTAS groupings
}
\keyword{datasets}
